@@ -13,10 +13,10 @@
     "bootstrap": "^5.1.3",
     "dayjs": "^1.11.3",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.6",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.8.0",
     "react-scripts": "5.0.1",
     "react-transition-group": "^4.4.2",
     "redux-persist": "^6.0.0",
@@ -55,7 +55,7 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
+    "@testing-library/react": "^14.0.0",
     "@types/jest": "^27.4.1",
     "@types/react-dom": "^17.0.13",
     "@types/react-router-dom": "^5.3.3",
