@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import axios from 'axios';
 import dayjs from 'dayjs';
 import 'dayjs/locale/ru';
@@ -14,6 +14,8 @@ import 'bootstrap/dist/css/bootstrap-grid.min.css';
 import 'antd/dist/antd.min.css';
 import { GlobalProvider } from './app/GlobalProvider';
 
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
 axios.defaults.baseURL = BASE_API_URL;
 axios.defaults.headers.authorization = `Bearer ${localStorage.getItem(LS_TOKEN_KEY)}`;
 
@@ -21,11 +23,8 @@ dayjs.extend(relativeTime);
 dayjs.extend(calendar);
 dayjs.locale('ru');
 
-ReactDOM.render(
+root.render(
   <GlobalProvider>
     <App />
-  </GlobalProvider>,
-  document.getElementById('root')
+  </GlobalProvider>
 );
-
-// fake
